refactor(ingest): centralize stored-procedure models in a pydantic models module
Move BaseProcedure and the procedure-call value object out of base.py /
lineage.py into a new stored_procedures/models.py, converting both from
dataclasses to pydantic v2 models:

- BaseProcedure gains extra="forbid" so a mistyped kwarg fails loudly.
- The formerly private _ProcedureCall becomes a public, frozen ProcedureCall;
  its `schema` field is renamed to `db_schema` (schema shadows a BaseModel
  attribute and now matches SchemaKey.db_schema).
- get_procedure_flow_name moves alongside the models it serves.

Repoint all importers (mysql, oracle, postgres, timescaledb, db2, doris, hana,
mssql, snowflake, sql_common) and the related tests at models.py. No import
cycle is introduced: models imports neither base nor lineage.

The pydantic validation surfaced a test-only bug where a bare Mock inspector
returned a Mock for denormalize_name(); the oracle test now returns a concrete
schema string, matching production behavior.
