+++ metadata-ingestion/src/datahub/ingestion/source/sql/oracle.py
@@ -71,8 +71,10 @@ from datahub.ingestion.source.sql.sql_utils import (
     get_domain_wu,
 )
 from datahub.ingestion.source.sql.stored_procedures.base import (
-    BaseProcedure,
     generate_procedure_workunits,
+)
+from datahub.ingestion.source.sql.stored_procedures.models import (
+    BaseProcedure,
     get_procedure_flow_name,
 )
 from datahub.ingestion.source.usage.usage_common import BaseUsageConfig
